test(react): add test cases for devMode in production mode, meet coverage 100%
