+++ .github/workflows/ci.yml
@@ -20,18 +20,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        command:
-          [
-            'ci:attw',
-            'ci:eslint',
-            'ci:lighthouse',
-            'ci:publint',
-            'ci:sherif',
-            'ci:type',
-            'test',
-            'test:production',
-            'build',
-          ]
+        command: ['ci:attw', 'ci:eslint', 'ci:lighthouse', 'ci:publint', 'ci:sherif', 'ci:type', 'test', 'build']
     steps:
       - uses: actions/checkout@v4
       - uses: ./.github/actions/pnpm-setup-node
