+++ .gitignore
@@ -37,3 +37,6 @@ graph/
 
 # lighthouse
 .lighthouseci/
+
+# test
+tsconfig.vitest-temp.json
