+++ packages/react/src/utils/getNull.spec.ts
@@ -1,0 +1,8 @@
+import { describe, expect, it } from 'vitest'
+import { getNull } from './getNull'
+
+describe('getNull', () => {
+  it('should be return null', () => {
+    expect(getNull()).toBeNull()
+  })
+})
