+++ packages/react/src/utils/getNull.ts
@@ -1,0 +1,1 @@
+export const getNull = () => null
