--- packages/react/vitest-production.config.ts
@@ -1,16 +1,0 @@
-import { defineConfig } from 'vitest/config'
-import packageJson from './package.json'
-
-export default defineConfig({
-  test: {
-    name: packageJson.name,
-    dir: './src',
-    include: ['**/*.production.*'],
-    environment: 'jsdom',
-    globals: true,
-    setupFiles: './vitest-production.setup.ts',
-    coverage: {
-      provider: 'istanbul',
-    },
-  },
-})
